Fix memory corruption issue with IPv6 support
We added IPv6 support to tun2socks in #271, but there was a
typo that would cause a double-free if IPv6 support were enabled
(which it was not).  This change fixes the typo so that IPv6
works properly.

This change does not yet include the rebuilt libtun2socks.so, due
to NDK compatibility issues.