Fix type mismatches in OutlineEvent.
A parameter of type (interface {}) => void can't take a function of type
(interface {param: type}) => void as an argument.  I'm not sure why,
this seems like a bug in Typescript's type system, but this seems common
enough that I'd expect there's an explanation.

Luckily, all OutlineEvents take a server as a parameter.  This creates
the issue, however, that we are trying to get the constructor of an
interface which only says it has a Server parameter -- not a constructor
which takes it.  So we create a ServerConstructible interface specifying
types which take Servers as parameters and have them publicly available,
and then get its constructor through Object.getPrototypeOf()

This error only triggered when directly invoking `tsc -p src/www`.  I'm
not sure why it didn't trigger on `yarn do src/www/build`, which is
basically the same thing.