Silence app version warning
When running the elctron app, a runtime error occurs due to no version
number being present in package.json.  This change temporarily injects a
version number into package.json and cleans up when we're done.

The process creates a ending newline in package.json, so we check it in
in order to keep that from polluting the git changes.

Also adds a note about running the electron app from the commandline in
README.md