Ignore Typescript files for Prettier
git-clang-format doesn't re-stage its diffs.  This means that if both clang-format and pretty-quick want to format a file, pretty-quick will error out complaining about a partially staged file left by clang-format.  This change tells pretty-quick to ignore our typescript files.

A better option which will require more churn would be to drop git-clang-format in preference for Prettier completely.  This gives the advantage that a separate formatting commit doesn't need to be made after the fact, or we don't need to remember to git commit --amend in order to include git-clang-format's diffs.

As a point of note, https://reviews.llvm.org/D41147 is an accepted, but not merged, change from 2017 which would add a --staged option to git-clang-format doing exactly what we want it to do.  Both klimek and djasper still work at Google C++ so it could be possible to get in touch with them to see if we can't pick up the last stretch and merge it into git-clang-format upstream.