➖ fix(deps): statically link vcruntime140.dll into smartdnsblock.exe (#1165)
Previously `smartdnsblock.exe` depends on `VCRUNTIME140.dll` which was not included in the electron installer, therefore some users might not be able to connect to a VPN server if they did not install any VC runtimes before. (following is the binary dump for the previous `smartdnsblock.exe`)

![image](https://user-images.githubusercontent.com/93548144/151441952-3b2b07d5-9ef3-4b97-acf8-16cf40bd0bec.png)

In this PR, I statically linked the VC runtime library into the `smartdnsblock.exe` itself. So the binary size increases a lot, but now the dependencies are much simpler:

![image](https://user-images.githubusercontent.com/93548144/151442104-24fd260f-0b19-46af-bf6f-38deb3770643.png)

Note the `smartdnsblock.exe` is currently a released version in x86 instead of a x64.

Fixes #1130