🚑️ hotfix(electron): fix cordova ReferenceError in electron (#1238)
In this PR, I fixed an issue found during the v1.7.0 release: `Uncaught ReferenceError: cordova is not defined`. This is because we are referencing a non-exist global variable in `cordova?.platformId ?? 'Electron'`. We can use a `global` or `windows` prefix like `global.cordova?.platformId`, but that might not work well either in electron or browser. Therefore, I chose to use a more traditional way of determining whether a global variable exists: `typeof cordova === 'undefined'`.

BTW, I also fixed the start action script, it was passing the wrong arguments to `build` action.