💊 fix(electron): fix wrong connection status when error is thrown during connect (#1286)
Previously when user clicks the "CONNECT" button, but encounters some errors (for example, routing daemon service is not running), our app still thinks we have *connected*, which is wrong. Even worse, when user tries to "DISCONNECT" then, our app just freezes forever.

In this PR, I fixed this issue by always making sure the `onceDisconnected` event is fired in `RoutingDaemon` when `stop()` method is called. It will then trigger the `onceDisconnected` event in `ShadowsocksLibevBadvpnTunnel`, and reset the state.