fix(electron): 💊 fix routing daemon installation issue in Linux (#1274)
In this PR, I fixed the routing daemon installation issue in Linux. Previously the logic was to launch `BadvpnTun2socks` first, and then tried to install daemon if it is not installed. This will cause failures because we did not relaunch `BadvpnTun2socks` after the installation. Now the new workflows are illustrated below.

#### Normal User Experience

https://user-images.githubusercontent.com/93548144/167020144-16273b08-9e0d-4029-b8f5-7b3731a89cee.mov

#### Edge Cases

https://user-images.githubusercontent.com/93548144/167020645-5d60e58e-1300-48ed-8161-e60479444f94.mov


Fixes #1269 .