refactor(electron): ♻️ use async model for ChildProcessHelper (#1455)
In this PR, I refactored `ChildProcessHelper` to use the async model when launching new child process, to replace the old event model. So the caller can get out the exit code more easily. As a side effect of this refactoring, I used `ChildProcessHelper` to implement the `checkConnectivity` feature in `GoVpnTunnel`, I also fixed #1449 by reading the exit code from `ChildProcessHelper`.