Propagating a Proxy Config [JSON] String from browser down to go-tun2socks layer
The config string is mostly opaque (except host retrieval in iOS) and passed along the layers unchanged.

Note: In iOS, We no longer update the config's host with the resolved IP address and pass along the original hostname, if not an IP already, to tun2socks layer so the config string effectively remains unchanged all the way.