feat(client): add callback/event mechanism between TypeScript and Go (#2330)
This PR introduces a `callback` mechanism that allows TypeScript code to subscribe to events triggered in the Go code.

To demonstrate the functionality, this PR implements a VPN connection state change event. The TypeScript code can now subscribe to this event and receive updates on the VPN connection state.

This mechanism can be leveraged for both electron and the Cordova code.