fix(client/linux): add timeout for CloseVPN (#2366)
This PR adds a 10-second timeout to `CloseVPN()`.  The `CloseVPN()` method can hang a very long time if there's outgoing traffic and the network is unreachable, due to lwIP holding a global lock.

#### Limitation

This is a workaround, so the user can quit the app. However, the lwIP lock *still* persist, requiring an **application restart**.